Start converting DateField to tsx
